CPU utilization as % is now available for Windows. get_cpu_times() now also raises NoSuchProcess in the event the process is nonexistent or disappears on us.
Process objects now initilize CPU time information for use with get_cpu_percent() in the constructor so we can avoid calling time.sleep() in the body of get_cpu_percent() for the first call. This may return 0.0% if the user calls it too soon after the initialization of the Process object but the alternative requires a forced time.sleep() in the function body which could have drawbacks in some circumstances.