Fix issue 209 (Windows - disk_partitions() options field).
We now provide a string for "options" value on Windows by including:
- "ro" or "rw" permissions
- the drive type (e.g. "fixed", "cdrom", "removable")

...as a single string with comma separated values, similarly to what happens on UNIX.

The "fstype" field is now filled with the real filesystem type (e.g. "NTFS" of "FAT").
In a sense, this breaks backward compatibility but I doubt anyone was actually relying on the previous value which was referring to the DRIVE TYPE and not the FILESYSTEM TYPE, and therefore it was misleading/incorrect in the first place.