Rewrite get_open_files to use a worker thread
get_open_files spawns a worker thread to process calls to
NtQueryObject().

If the call timesout (100ms), the thread is terminated and the file
handle is skipped.  The ConvertThreadToFiber call prevents the initial
stack from leaking when the thread is terminated on Windows XP/2003.

The code has been cleaned up and struct redefinitions in ntextapi have
been removed and with conflicts using #define's.