Support an edge case where XP sets dwLength to 0
There's an edge case on Windows XP where an infinite loop is possible.

A call to NtQueryObject with a NULL buffer and a buffer size of 0 will
return STATUS_INFO_LENGTH_MISMATCH but dwLength will be 0.

This change will pre-allocate a buffer of size (MAX_PATH + 1) *
sizeof(WCHAR) and skips the handle if the required dwLenght is 0.