swap: fix finding the 'swap' utility on SunOS
This is a bit of a kludge, but allows to find the swap utility when
invoced by normal users which shouldn't have the sbin paths in PATH by
default.  One can argue the paths should be appended instead of
prepended, but I doubt whether one wants to use another swap utility,
since the code relies on its output format.