TestUnicode: Instead of using Python's executable, compile one if possible
Some environments don't support running 'python' outside of its directory.
There is already a test that compiles an executable for testing - so we use
that code if possible, and fall back to copying the Python executable only if
gcc isn't available.

This also allows test_prog_w_funky_name to be tested on Winows, where
moving the executable is ok (at least for default Python distributions).