First pass at implementing Process.create_time for Cygwin
I left this unimplemented originally just because I wanted to take the
time to understand what I was doing before continuing.  Turns out it can
be implemented almost the same as on Linux, except that the process
start time is in the 22nd field of /proc/<pid>/stat instead of the 21st.
I can't find enough documentation for Cygwin's /proc filesystem, but I
confirmed this in the source.  Other than that the semantics are the
same as on Linux.

So much of this was copied agained from the Linux module, but I will
refactor later.