Compile at least the arch/windows/process_info.c module into the cygwin module so that the functions in it can be reused there.
This required some tweaks to ntextapi.h, since a few of the
types defined there are already included in the winimpl.h included
with Cygwin.

This includes are own implementation of PyErr_SetFromWindowsErr which
is not defined in Python for Cygwin.

TODO: Because this PyErr_SetFromWindowsErr is lifted, only slightly
modified, from the Python source it should include a copyright notice
to the effect.  This implementation probably does not work quite right
on Python 2 either--it should return an actual WindowsError there.