Initial implementation of net_if_addrs() for the cygwin module.
The only obvious issue currently is that rather than the human-friendly interface
names returned by this function in the normal Windows implementation, the
interface names are instead some UUID looking things for each interface (the same
as appear in the registry in various places to represent the interface).  This is
fine for now, but not ideal.  Should be able to map these to friendlier names.