Very rough first pass at implementing net_if_stats for Cygwin
This is mostly copied from the Windows module with a few small tweaks, but compiling this
is a bit of a pain and somewhat fragile.  Will have to come back to this for a closer look
later, but for now I just want to get all of the API implemented as best I can.