After researching the subject extensively, I've determined it's not possible to enumerate UNIX sockets on Cygwin
Cygwin emulates UNIX sockets as a INET socket.  Although it does list them in /proc/<pid>/fd as some 'socket:[N]' like on Linux, the 'N' here is only of internal meaning to Cygwin (this is true on Linux as well), but there is nothing resembling /proc/net/unix or anything else that allows one to map that number N to an actual socket file (such a file *does* exist, and in theory we could find it be enumerating ALL files on the system until it is found, but this is not practical).  I can't find any other obvious way to do that.  The only practical solution would be if Cygwin had an equivalent to /proc/net/unix.