Turns out the process creation time from /proc/<pid>/stat cannot be relied on in Cygwin, due to a rounding error there which can cause it to return fluctuating results over time.
This is really bad because process creation time is part of the
'identity' psutil util uses for a process when comparing it for equality
to another process.

This once again copies code from the windows module into the cygwin
module, which I intend to refactor later.  I also plan to address the
upstream bug in Cygwin, but until that's fixed this workaround is
needed.