re #1040: bif one win/ unicode / memory_maps()
Fix memory_maps() which was returning an invalid encoded path in case of
non ASCII path on both Python 2 and 3. Use GetMappedFileNameW instead of GetMappedFilenameA in
order to ask the system an actual unicode path. Also, on Windows
encode unicode back to str/bytes by using default fs-encoding + "replace"
error handler.
This paves the way for fixing other APIs in an identical manner.