fix valid_types in memory_percent
pfullmem._fields is always added twice to valid_types so
the message about invalid memtype lists the types twice too.
pfullmem is available on all platforms and is always the
same as or a superset of pmem. We can look at its fields
only to get all valid_types.
Also we can check whether to use memory_full_info or not
by checking the fields of pfullmem vs. pmem instead of
using hard coded mem types.