Small changes (#1158)
* fix valid_types in memory_percent

pfullmem._fields is always added twice to valid_types so
the message about invalid memtype lists the types twice too.
pfullmem is available on all platforms and is always the
same as or a superset of pmem. We can look at its fields
only to get all valid_types.
Also we can check whether to use memory_full_info or not
by checking the fields of pfullmem vs. pmem instead of
using hard coded mem types.

* remove workaround made for Solaris on AIX

The problem described in the comment doesn't apply for AIX

* update "oneshot" table in documentation

* Removed "nice" and "ionice" which are not boosted
* Removed "~Process." prefix which was only on a few
methods and not others
* Added AIX

Fixes #1157

* small AIX additions to docs