Replaced nice with priority on Windows
Windows uses a priroity like 'Normal' to communicate a process's
niceness rather than a number. When running top.py on Windows, the
niceness was coming through as a Priority enum which was cast to
a string like 'Priority.NORMAL_PRIORITY_CLASS', completely destroying
the column widths.

This commit addresses the issue like so:
  - A new function 'niceness2priority' was added which will return a
    more output-firendly word like 'Normal' or 'Idle' on Windows and
    the niceness on Linux.
  - The NI column will become the PRI column on Windows.