Fixed Ctrl+C behavior on Windows (again)
Commit 339349eb failed to allow 'q' and Ctrl+C to exit on Windows in
new terminals. Once the python process was terminated, subsequent
executions of the application were able to exit on user input as
expected, which is why it was missed in testing. The issue did not
affect Linux.

This commit adds better 'q' and Ctrl+C to quit support for Windows
and Linux.