Fix corner case when /etc/mtab doesn't exist and procfs=/proc
In some Linux configurations the `/etc/mtab` does not exist but the
procfs_path is equal to `/proc`.
With the fix done for the issue #1307, the described configuration didn't work.

This Commit introduce an additional check that verifies if the
`/etc/mtab` file exists before using it, else it defaults to `<procfs_path>/self/mounts`

Signed-off-by: cedric lamoriniere <cedric.lamoriniere@datadoghq.com>