NetBSD fixes (#1526)
* Fix psutil_get_cmd_args() for NetBSD

Do not overallocate the buffer prompting for KERN_ARGMAX.
It also fixes the code as KERN_ARGMAX was received into a size_t
type, while kernel returned int. It caused argmax to contain
garbage and allocation was randomly crashing with new jemalloc in
the basesystem.

New code prompts for exact buffer size before allocation and
stores argv[] inside it.

Bug investigated by Leonardo Taccari.

* Stop including sys/user.h for NetBSD

This header was removed from the OS as it was empty.