Add test demonstrating net namespace invisibility
Added test that opens two processes. One behaves as if it's in the same
network namespace as the testing process, the other - as if it's another
network namespace. Psutil should detect a connection from the first but
not from the second process before the fix. After it should detect from
both.