Add a resuable psutil_getpagesize() utility
Related to #1885. getpagesize() is deprecated on macOS, POSIX.1-2001 and
possibly other UNIXes.

The problem emerged on macOS but getpagesize() is also used in FreeBSD,
NetBSD, OpenBSD and AIX, so it makes sense to do this in one place only,
similarly to Windows which also provide a psutil_getpagesize() utility.

Follow cPython's mmapmodule.c and resourcemodule.c lead and rely on
`sysconf(_SC_PAGESIZE)` instead.

Leave getpagesize() in place as last resort/attemp for systems where it's
not deprecated or they still legitimately rely on it.

Also provide a python wrapper so we can test the return value of this C
function against Python's stdlib.

Signed-off-by: Giampaolo Rodola <g.rodola@gmail.com>