Resolve race condition in Process.threads() (#2151)
* Resolve race condition in Process.threads()

Process.threads() has a race condition triggered when a thread exits
after the open_binary() call and before the read() call. When this
happens, the read() call raises ProcessLookupError.

Handle the race condition by catching ProcessLookupError from read() and
treating the same as a FileNotFoundError from open_binary(). This is the
same approach used in ppid_map().

Signed-off-by: Daniel Li <daniel.li@deshaw.com>

* Also catch ProcessLookupError in open_files()

Signed-off-by: Daniel Li <daniel.li@deshaw.com>