fix: race condition in test_posix.TestProcess.test_cmdline
On macOS the official python installer exposes a python wrapper that
executes a python executable hidden inside an application bundle inside
the Python framework.

There's a race condition in in test_posix.TestProcess.test_cmdline between
the ps call & the psutil call depending on the completion of the execve
call so let's retry on failure for this test.

Signed-off-by: mayeut <mayeut@users.noreply.github.com>