Exclude external batteries from sensors_battery
The single check for battery in power_supply was causing sometimes to grab external device battery statuses since they have the form of hid_<mac>_battery, which match the bat filter. 

However they do not have an ACPI path to the system itself. 

This takes into account that detail and properly exclude them from being included as an internal battery.