switch to src layout
This switches the project to use a src/ layout, keeping tests in the top
level. See for more information:

https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/

One of the benefits of this layout is that tests are now in the top
level directory. They are no longer part of the generated wheel, keeping
dist size small and installations clean.

See also https://github.com/giampaolo/psutil/pull/2630