test_process_all.py: enforce the "fork" multiprocessing method (#2666)
There are three method for launching processes in the multiprocessing module. "fork" is the only one that does not
launch an additional "resource_tracker" process. This process ignores SIGINT and SIGTERM, so that it is not ended by "reap_children", which messes other tests (and may lead to hanging).
Fixes #2663