BSD / OSX: increase chances to recognize zombie processes (#2672)
`psutil_raise_for_pid()` is a utility function which we used for those syscalls that do not return an error that we can directly translate into an exception. E.g. they don't even set errno. So we have to guess. 

Let's increase this guessing strategy, and check if PID is zombie from within `psutil_raise_for_pid()` itself. This will increase the chances for psutil to raise `ZombieProcess` instead of `AccessDenied`, `OSError` or `RuntimeError`.

`psutil_raise_for_pid()` is used by a lot of public process APIs on OSX and BSD. A non extensive list include (but there are more):

- open_files()
- cwd()
- num_fds()
- memory_maps()
- net_connections()
- exe()