Make `psutil.test.MemoryLeakTestCase` public (#2682)
This function has been used internally for a long time to check memory leaks in all psutil C functions, but it's generic enough that it's worth it to make it public and officially document it. MemoryLeakTestCase now lives under the new psutil.test namespace, which only defines this class.

The leak detection logic now considers VMS, RSS, and USS instead of only USS, making it more reliable across different platforms and allocator behaviors.

Next step is gonna be using also mallinfo on Linux (see #1275) to make it even more reliable.